Create a splash screen and main menu system
Create a splash screen with the team's logo under splash_screen.tscn. The interact button is pressed to transition into the main menu.

Create a transition_screen.tscn node which creates a "fade in" or "fade out" effect. A transition_ended signal is emitted by this node when a transition is completed.
This can be reused in different scenes later on.

Create a main menu with a texture background and a selection pane. The selection pane is controlled using the up and down arrow keys with an outline around a focused element.
When a label is interacted with, the button_press function is called on that label.

Add the DataManager node to the main menu to clear save data when a new game is started and to verify whether data exists. If no data exists, then the load data label will be faded out.