Create a save data manager
Create DataManager, a save data manager which can be loaded as a Node. This class contains functions to save data, load data, and get whether any data exists.
Documentation is included in the class as well.

The location of the saved data and the file extensions used for data are stored as constants in the data_manager.gd file.

Data is saved as a JSON file with lossy type conversions fixed when the JSON file is loaded. A backup file will be created periodically or on the first save from when the game starts.