Load start position from save file
Place the player based on the saved position in the save data.

The starting location on the map is loaded from MAP_POSITION_KEY in the save data. If this key does not exist, then the location at the house is used.

Reset the current save file to a blank dictionary when a new game is started. This resets the starting location to the house.