Create maps and fix spawning and dialogue
Create separate scenes for each of the three maps using updated tiles. Each tile is now 50 px by 50 px. Decoration has also been added to the maps so that they are not blank.

The player is now spawned by placing them at a set spawn location Node2D in a map.

Add background music which autoplays when the scene is loaded.

Fix an interaction bug with dialogue where finished dialogue could not be skipped, and fix a JSON parsing bug where a null texture would cause an error.